Ignore already-seen statuses
This prevents the bot from getting obsessed with a word, which is
especially important because people tend to interact with it in a
flurry.

The already-seen statuses are **not** serialized into the bot status,
they are only kept in-memory. For most purposes this is enough, to the
point that I might even consider clearing the HashSet from memory on
occasion.